windows_collector/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncclick==8.1.8.0
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
gensim==3.8.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mashumaro==3.15
multidict==6.3.2
numpy==2.2.4
pandas==2.2.3
propcache==0.3.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-kasa==0.7.7
pytz==2025.2
requests==2.32.3
scipy==1.15.2
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
style==1.1.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
update==0.0.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
